Extract createDynamicFilterEvaluator to DynamicFiltersTestUtil
